Separate content and file
